+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Include object ID in "local object" error message.
 - Use mitra-web background color on OAuth pages.
+- Delete orphaned local groups after their owner is deleted.
 
 ### Deprecated